Moved Layout into a component folder, and moved Nav and Footer into that folder
