+++ src/components/Layout/index.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom'
-import Footer from './Footer/Footer'
-import Nav from './Nav/Nav'
+import Footer from './Footer'
+import Nav from './Nav'
 
 export default function Layout() {
   return (
